LayoutTests/media/modern-media-controls/play-pause-button/play-pause-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/play-pause-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlayPauseButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const playButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("playButton.element.localName", "button");
shouldBeTrue('playButton.element.classList.contains("icon")');
shouldBeTrue('playButton.element.classList.contains("play-pause")');
shouldBeEqualToString("playButton.iconName", `${Icons.Play}`);
shouldBeFalse("playButton.playing");

const fullscreenPlayButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
const iOSPlayButton = new PlayPauseButton({ layoutTraits: LayoutTraits.iOS });

const pauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS });
pauseButton.playing = true;
shouldBeTrue("pauseButton.playing");
shouldBeEqualToString("pauseButton.iconName", `${Icons.Pause}`);

const fullscreenPauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
fullscreenPauseButton.playing = true;

const iOSPauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.iOS });
iOSPauseButton.playing = true;

scheduler.frameDidFire = function()
{
    if ([playButton, pauseButton, iOSPlayButton, iOSPauseButton, fullscreenPlayButton, fullscreenPauseButton].some(button => button.width == 0))
        return;

    shouldBeTrue('playButton.element.style.webkitMaskImage.includes("macOS/play@")');
    shouldBeTrue('pauseButton.element.style.webkitMaskImage.includes("macOS/pause@")');
    shouldBeTrue('iOSPlayButton.element.style.webkitMaskImage.includes("iOS/play@")');
    shouldBeTrue('iOSPauseButton.element.style.webkitMaskImage.includes("iOS/pause@")');
    shouldBeTrue('fullscreenPlayButton.element.style.webkitMaskImage.includes("macOS/play-fullscreen@")');
    shouldBeTrue('fullscreenPauseButton.element.style.webkitMaskImage.includes("macOS/pause-fullscreen@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
